+++ src/bootstrap.ts
@@ -1,0 +1,23 @@
+export function checkReady() {
+  const DEVICE_READY_TIMEOUT = 5000;
+
+  declare var window;
+
+  // To help developers using cordova, we listen for the device ready event and
+  // log an error if it didn't fire in a reasonable amount of time. Generally,
+  // when this happens, developers should remove and reinstall plugins, since
+  // an inconsistent plugin is often the culprit.
+  const before = Date.now();
+
+  let didFireReady = false;
+  document.addEventListener('deviceready', () => {
+    console.log('DEVICE READY FIRED AFTER', (Date.now() - before), 'ms');
+    didFireReady = true;
+  });
+
+  setTimeout(() => {
+    if (!didFireReady && window.cordova) {
+      console.warn(`Native: deviceready did not fire within ${DEVICE_READY_TIMEOUT}ms. This can happen when plugins are in an inconsistent state. Try removing plugins from plugins/ and reinstalling them.`);
+    }
+  }, DEVICE_READY_TIMEOUT);
+}
